requirements.txt
astroid==2.2.5
autopep8==1.4.4
bandit==1.6.2
Click==7.0
colorama==0.4.1
entrypoints==0.3
falcon==2.0.0
flake8==3.7.8
flake8-polyfill==1.0.2
Flask==1.1.1
future==0.17.1
gitdb2==2.0.5
GitPython==3.0.2
gunicorn==19.9.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
mando==0.6.4
MarkupSafe==1.1.1
mccabe==0.6.1
pbr==5.4.3
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
PyYAML==5.1.2
radon==3.0.3
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
six==1.12.0
smmap2==2.0.5
SQLAlchemy==1.3.8
stevedore==1.31.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
